chore(backend): wip journey ios
