+++ backend/api/measure/app.go
@@ -18,6 +18,7 @@ import (
 	"backend/api/journey"
 	"backend/api/metrics"
 	"backend/api/paginate"
+	"backend/api/platform"
 	"backend/api/replay"
 	"backend/api/server"
 	"backend/api/span"
@@ -910,25 +911,52 @@ func (a App) GetLaunchMetrics(ctx context.Context, af *filter.AppFilter) (launch
 // and issue events involved in forming
 // an implicit navigational journey.
 func (a App) getJourneyEvents(ctx context.Context, af *filter.AppFilter, opts filter.JourneyOpts) (events []event.EventField, err error) {
-	whereVals := []any{
-		event.TypeLifecycleActivity,
-		[]string{
-			event.LifecycleActivityTypeCreated,
-			event.LifecycleActivityTypeResumed,
-		},
-		event.TypeLifecycleFragment,
-		[]string{
-			event.LifecycleFragmentTypeAttached,
-			event.LifecycleFragmentTypeResumed,
-		},
+	whereVals := []any{}
+
+	switch a.Platform {
+	case platform.Android:
+		whereVals = append(
+			whereVals,
+			event.TypeLifecycleActivity,
+			[]string{
+				event.LifecycleActivityTypeCreated,
+				event.LifecycleActivityTypeResumed,
+			},
+			event.TypeLifecycleFragment,
+			[]string{
+				event.LifecycleFragmentTypeAttached,
+				event.LifecycleFragmentTypeResumed,
+			},
+		)
+	case platform.IOS:
+		whereVals = append(
+			whereVals,
+			event.TypeLifecycleViewController,
+			[]string{
+				event.LifecycleViewControllerTypeViewDidLoad,
+				event.LifecycleViewControllerTypeViewDidAppear,
+			},
+			event.TypeLifecycleSwiftUI,
+			[]string{
+				event.LifecycleSwiftUITypeOnAppear,
+			},
+		)
 	}
 
 	if opts.All {
-		whereVals = append(whereVals, event.TypeException, false, event.TypeANR)
+		switch a.Platform {
+		case platform.Android:
+			whereVals = append(whereVals, event.TypeException, false, event.TypeANR)
+		case platform.IOS:
+			whereVals = append(whereVals, event.TypeException, false)
+		}
 	} else if opts.Exceptions {
 		whereVals = append(whereVals, event.TypeException, false)
 	} else if opts.ANRs {
-		whereVals = append(whereVals, event.TypeANR)
+		switch a.Platform {
+		case platform.Android:
+			whereVals = append(whereVals, event.TypeANR)
+		}
 	}
 
 	stmt := sqlf.
@@ -937,14 +965,25 @@ func (a App) getJourneyEvents(ctx context.Context, af *filter.AppFilter, opts fi
 		Select(`toString(type)`).
 		Select(`timestamp`).
 		Select(`session_id`).
-		Select(`toString(lifecycle_activity.type)`).
-		Select(`toString(lifecycle_activity.class_name)`).
-		Select(`toString(lifecycle_fragment.type)`).
-		Select(`toString(lifecycle_fragment.class_name)`).
-		Select(`toString(lifecycle_fragment.parent_activity)`).
-		Select(`toString(lifecycle_fragment.parent_fragment)`).
-		Where(`app_id = ?`, a.ID).
+		Where(`app_id = toUUID(?)`, a.ID).
 		Where("`timestamp` >= ? and `timestamp` <= ?", af.From, af.To)
+
+	switch a.Platform {
+	case platform.Android:
+		stmt.
+			Select(`toString(lifecycle_activity.type)`).
+			Select(`toString(lifecycle_activity.class_name)`).
+			Select(`toString(lifecycle_fragment.type)`).
+			Select(`toString(lifecycle_fragment.class_name)`).
+			Select(`toString(lifecycle_fragment.parent_activity)`).
+			Select(`toString(lifecycle_fragment.parent_fragment)`)
+	case platform.IOS:
+		stmt.
+			Select(`toString(lifecycle_view_controller.type)`).
+			Select(`toString(lifecycle_view_controller.class_name)`).
+			Select(`toString(lifecycle_swift_ui.type)`).
+			Select(`toString(lifecycle_swift_ui.class_name)`)
+	}
 
 	if len(af.Versions) > 0 {
 		stmt.Where("`attribute.app_version` in ?", af.Versions)
@@ -955,11 +994,19 @@ func (a App) getJourneyEvents(ctx context.Context, af *filter.AppFilter, opts fi
 	}
 
 	if opts.All {
-		stmt.Where("((type = ? and `lifecycle_activity.type` in ?) or (type = ? and `lifecycle_fragment.type` in ?) or ((type = ? and `exception.handled` = ?) or type = ?))", whereVals...)
+		switch a.Platform {
+		case platform.Android:
+			stmt.Where("((type = ? and `lifecycle_activity.type` in ?) or (type = ? and `lifecycle_fragment.type` in ?) or ((type = ? and `exception.handled` = ?) or type = ?))", whereVals...)
+		case platform.IOS:
+			stmt.Where("((type = ? and `lifecycle_view_controller.type` in ?) or (type = ? and `lifecycle_swift_ui.type` in ?) or (type = ? and `exception.handled` = ?))", whereVals...)
+		}
 	} else if opts.Exceptions {
 		stmt.Where("((type = ? and `lifecycle_activity.type` in ?) or (type = ? and `lifecycle_fragment.type` in ?) or (type = ? and `exception.handled` = ?))", whereVals...)
 	} else if opts.ANRs {
-		stmt.Where("((type = ? and `lifecycle_activity.type` in ?) or (type = ? and `lifecycle_fragment.type` in ?) or (type = ?))", whereVals...)
+		switch a.Platform {
+		case platform.Android:
+			stmt.Where("((type = ? and `lifecycle_activity.type` in ?) or (type = ? and `lifecycle_fragment.type` in ?) or (type = ?))", whereVals...)
+		}
 	}
 
 	if len(af.OsNames) > 0 {
@@ -1015,18 +1062,37 @@ func (a App) getJourneyEvents(ctx context.Context, af *filter.AppFilter, opts fi
 		var lifecycleFragmentClassName string
 		var lifecycleFragmentParentActivity string
 		var lifecycleFragmentParentFragment string
+		var lifecycleViewControllerType string
+		var lifecycleViewControllerClassName string
+		var lifecycleSwiftUIType string
+		var lifecycleSwiftUIClassName string
 
 		dest := []any{
 			&ev.ID,
 			&ev.Type,
 			&ev.Timestamp,
 			&ev.SessionID,
-			&lifecycleActivityType,
-			&lifecycleActivityClassName,
-			&lifecycleFragmentType,
-			&lifecycleFragmentClassName,
-			&lifecycleFragmentParentActivity,
-			&lifecycleFragmentParentFragment,
+		}
+
+		switch a.Platform {
+		case platform.Android:
+			dest = append(
+				dest,
+				&lifecycleActivityType,
+				&lifecycleActivityClassName,
+				&lifecycleFragmentType,
+				&lifecycleFragmentClassName,
+				&lifecycleFragmentParentActivity,
+				&lifecycleFragmentParentFragment,
+			)
+		case platform.IOS:
+			dest = append(
+				dest,
+				&lifecycleViewControllerType,
+				&lifecycleViewControllerClassName,
+				&lifecycleSwiftUIType,
+				&lifecycleSwiftUIClassName,
+			)
 		}
 
 		if err := rows.Scan(dest...); err != nil {
@@ -1044,6 +1110,16 @@ func (a App) getJourneyEvents(ctx context.Context, af *filter.AppFilter, opts fi
 				ClassName:      lifecycleFragmentClassName,
 				ParentActivity: lifecycleFragmentParentActivity,
 				ParentFragment: lifecycleFragmentParentFragment,
+			}
+		} else if ev.IsLifecycleViewController() {
+			ev.LifecycleViewController = &event.LifecycleViewController{
+				Type:      lifecycleViewControllerType,
+				ClassName: lifecycleViewControllerClassName,
+			}
+		} else if ev.IsLifecycleSwiftUI() {
+			ev.LifecycleSwiftUI = &event.LifecycleSwiftUI{
+				Type:      lifecycleSwiftUIType,
+				ClassName: lifecycleSwiftUIClassName,
 			}
 		} else if ev.IsException() {
 			ev.Exception = &event.Exception{}
@@ -1924,17 +2000,25 @@ func GetAppJourney(c *gin.Context) {
 		ID: &id,
 	}
 
-	team, err := app.getTeam(ctx)
-	if err != nil {
-		msg := "failed to get team from app id"
+	if err := app.Populate(ctx); err != nil {
+		msg := `failed to fetch app details`
 		fmt.Println(msg, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": msg})
+		status := http.StatusInternalServerError
+
+		if errors.Is(err, pgx.ErrNoRows) {
+			status = http.StatusNotFound
+			msg = fmt.Sprintf(`app with id %q does not exist`, app.ID)
+		}
+
+		c.JSON(status, gin.H{
+			"error": msg,
+		})
+
 		return
 	}
-	if team == nil {
-		msg := fmt.Sprintf("no team exists for app [%s]", app.ID)
-		c.JSON(http.StatusBadRequest, gin.H{"error": msg})
-		return
+
+	team := &Team{
+		ID: &app.TeamId,
 	}
 
 	userId := c.GetString("userId")
@@ -1973,13 +2057,15 @@ func GetAppJourney(c *gin.Context) {
 		return
 	}
 
+	fmt.Println("count journey events", len(journeyEvents))
+
 	var issueEvents []event.EventField
 
 	for i := range journeyEvents {
 		if journeyEvents[i].IsUnhandledException() {
 			issueEvents = append(issueEvents, journeyEvents[i])
 		}
-		if journeyEvents[i].IsANR() {
+		if app.Platform == platform.Android && journeyEvents[i].IsANR() {
 			issueEvents = append(issueEvents, journeyEvents[i])
 		}
 	}
